Splitting generic.service into login.service and user.service
