+++ client/src/app/auth/auth-guard.service.ts
@@ -6,11 +6,11 @@ import {
     CanActivateChild
 } from '@angular/router';
 
-import {GenericService}      from '../generic.service';
+import {LoginService}      from '../remote/login.service';
 
 @Injectable()
 export class AuthGuard implements CanActivate, CanActivateChild {
-    constructor(private genericService: GenericService, private router: Router) {
+    constructor(private loginService: LoginService, private router: Router) {
     }
 
     canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): boolean {
@@ -23,11 +23,11 @@ export class AuthGuard implements CanActivate, CanActivateChild {
     }
 
     checkLogin(url: string): boolean {
-        if (this.genericService.loggedIn()) {
+        if (this.loginService.loggedIn()) {
             return true;
         }
         // Store the attempted URL for redirecting
-        this.genericService.redirectUrl = url;
+        this.loginService.redirectUrl = url;
         // Navigate to the login page
         this.router.navigate(['/login']);
         return false;
