+++ client/src/app/remote/in-memory-data.service.ts
@@ -1,7 +1,7 @@
 import {Injectable} from '@angular/core';
 import {InMemoryDbService} from 'angular-in-memory-web-api';
 
-import {User} from './user';
+import {User} from '../user';
 
 @Injectable()
 export class InMemoryDataService implements InMemoryDbService {
